Revert "Updating"

+++ lldb/test/API/types/AbstractBase.py
@@ -34,11 +34,14 @@ class GenericTester(TestBase):
         # used for all the test cases.
         self.exe_name = self.testMethodName
         golden = "{}-golden-output.txt".format(self.testMethodName)
-        self.golden_filename = self.getBuildArtifact(golden)
+        if configuration.is_reproducer():
+            self.golden_filename = self.getReproducerArtifact(golden)
+        else:
+            self.golden_filename = self.getBuildArtifact(golden)
 
     def tearDown(self):
         """Cleanup the test byproducts."""
-        if os.path.exists(self.golden_filename):
+        if os.path.exists(self.golden_filename) and not configuration.is_reproducer():
             os.remove(self.golden_filename)
         TestBase.tearDown(self)
 
@@ -101,6 +104,9 @@ class GenericTester(TestBase):
             # copy remote_path to local host
             self.runCmd('platform get-file {remote} "{local}"'.format(
                 remote=remote_path, local=self.golden_filename))
+        elif configuration.is_reproducer_replay():
+            # Don't overwrite the golden file generated at capture time.
+            self.runCmd('process launch')
         else:
             self.runCmd(
                 'process launch -o "{local}"'.format(local=self.golden_filename))
